Add option to allow user to select keys in response to request identities.
This works in Agent mode only. When a client program requests a list of
available keys, if more than one key is available a dialog is shown to allow
the user to select which key(s) that should be returned. This is useful when
the user has a large number of keys. For example, openssh server only allows
6 keys to be returned by default.

Issue dlech/KeeAgent#38.