Separate public key from private key.
We were trying to let the public key piggy-back on the private key by
replacing any public key included in a private key with one from a
separate public key file. However this proved to be problematic since
we weren't consistently replacing the public in all cases.

This separates the public key loading from the private key so that we
end up with more consistent results.

Fixes #369.