DecryptProgressDialog: fix crash
System.Progress calls the callback on a different thread, so there is a race condition between the background worker completing and the final progress callback being called. If the callback is called too late, we get an InvalidOperationException when calling ReportProgress().

To fix it, we create our own IProgress implementation that calls ReportProgress() syncronously.

Fixes: https://github.com/dlech/KeeAgent/issues/366